// src/data/graphql/models/profileGridViewCustomisation/initCommunityProfilePage.ts
export enum ProfilePageType {
  grid = 'grid',
}
export type InitCommunityProfilePageMutationResponse = {
  initCommunityProfilePage: {
    __typename: string;
    message: string;
    reference: {
      persona: {
        persona_id: string;
        user_handle: string;
        profile_name: string;
        profile_avatar_url: string;
        profile_bio: string;
        own_referral_code: string;
      };
      community_id: string;
      page_id: string;
      page_name: string;
      page_icon_url: string;
      seo_metadata: Record<string, unknown>;
      is_private: boolean;
      design: {
        links: {
          id: string;
          title: string;
          icon: string;
          url: string;
          tracking_url: string;
        }[];
        attributes: {
          badgeMetadata: {
            point_id: string;
            cta: string;
            thumbnail_url: string;
            img_url: string;
            display_name: string;
            free_claim: boolean;
            community_id: string;
          };
          balance: number;
          communityId: string;
          personaId: string;
          pointId: string;
        }[];
      };
      layout: {
        links: {
          id: string;
          title: string;
          icon: string;
          url: string;
          tracking_url: string;
        }[];
        attributes: {
          badgeMetadata: {
            point_id: string;
            cta: string;
            thumbnail_url: string;
            img_url: string;
            display_name: string;
            free_claim: boolean;
            community_id: string;
          };
          balance: number;
          communityId: string;
          personaId: string;
          pointId: string;
        }[];
      };
      page_config: Record<string, unknown>;
      generated_url: string;
      page_type: ProfilePageType | string;
      page_published: boolean;
      created_at: number;
      updated_at: number;
    };
    status: string;
  } & GraphqlError;
};

export const INIT_COMMUNITY_PROFILE_PAGE_MUTATION = `
mutation initCommunityProfilePage {
  initCommunityProfilePage(
    communityId: "1770071e-0000-0000-0000-1770071e7000"
  ) {
    ... on SimpleMutationResponse {
      __typename
      message
      reference
      status
    }
    ... on GraphqlError {
      __typename
      error
      message
      statusCode
    }
  }
}
`;